requirements.txt
python-telegram-bot==20.6
Flask==2.3.3
gspread==5.12.0
oauth2client==4.1.3
openai>=1.0.0
nest_asyncio